Add a frontend web service and test-data population
docker-compose.yaml is extended to have a frontend service running
on localhost:7080/localhost:7443 that proxies to the registry and
index.

registry_copy.py is a Python script that handles copying Docker/OCI
manifests and manifest-lists to and from registries. It's used to
populate test data when the frontend is started.

Test data is abstracted so you can run:

 TEST_DATA=<some_dir> docker-compose up

And a contents file describes what to populate the registery with.