Embed version information into the binary
Include information about the current Git version and build time
in the binary. This is displayed in a startup message and used to
generate a "BuildId" - 10 hex digits that should change when
the source code changes. (It has a timestamp in the hash contents
when the local git tree is dirty, but not otherwise.)