Split /index into /index/static and /index/dynamic
Make clients declare an intent - is the query they are making a fixed query
that will be made repeatedly or is is dynamically created based on user
input?

 /index/static - caching is probably useful
 /index/dynamic - caching is unlikely to be useful

The idea of this is to allow a frontend to cache static queries without
having dynamic queries explode the cache size. For dynamic queries,
Cache-Control: no-cache is set in the returned headers.